+++ src/Events/Input.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <glm/vec2.hpp>
+#include <glm/ext/vector_int2.hpp>
 
 #include "Keycodes.hpp"
 #include "MouseCodes.hpp"
@@ -17,7 +17,7 @@ class Input {
     bool IsKeyPressed(KeyCode);
     bool IsMouseButtonPressed(MouseCode);
 
-    glm::vec2 GetMousePosition();
+    glm::ivec2 GetMousePosition();
 
    private:
     Window& m_Window;
